requirements.txt
python-dotenv
pandas
numpy
matplotlib
plotly
tabulate
chainlit==1.0.101
langchain==0.1.0
langchain-community
langchain-openai
langchain-experimental
langchainhub
transformers
torch
openai
#ollama==0.1.0
#httpx>=0.25.2,<0.26.0
yfinance
moment
forex-python==1.6
wikipedia
duckduckgo-search
numexpr
chromadb
pymupdf